// resources/js/types/pagination.ts
export interface PaginationLink {
    url: string | null;
    label: string;
    page: number;
    active: boolean;
}

export interface PaginationMeta<T> {
    current_page: number;
    data: T[];
    first_page_url: string;
    from: number;
    last_page: number;
    last_page_url: string;
    links: PaginationLink[];
    next_page_url: string | null;
    path: string;
    per_page: number;
    prev_page_url: string | null;
    to: number;
    total: number;
}